requirements.txt
gradio==5.29.0
fastapi
uvicorn==0.34.0
transformers
torch
sentencepiece==0.2.0
gtts
python-dotenv==1.1.0
